fix: refactor and resolve errors
