+++ src/resources/boards/boards.service.ts
@@ -1,19 +1,18 @@
-import * as boardRepo from './boards.memory.repository';
+import * as boardsRepo from './boards.memory.repository';
 import { IBoard } from './boards.model';
+import * as tasksService from '../tasks/tasks.service';
 
-const tasksService = require('../tasks/tasks.service');
+const getAll = () => boardsRepo.getAllBoards();
 
-const getAll = () => boardRepo.getAllBoards();
+const getBoard = (id: string) => boardsRepo.getBoard(id);
 
-const getBoard = (id: string) => boardRepo.getBoard(id);
+const createBoard = (board:IBoard) => boardsRepo.createBoard(board);
 
-const createBoard = (board:IBoard) => boardRepo.createBoard(board);
-
-const updateBoard = (id:string, data:IBoard) => boardRepo.updateBoard(id, data);
+const updateBoard = (id:string, data:IBoard) => boardsRepo.updateBoard(id, data);
 
 const deleteBoard = (id:string) => {
   tasksService.deleteTasksFromBoard(id);
-  return boardRepo.deleteBoard(id);
+  return boardsRepo.deleteBoard(id);
 };
 
 export {
